public-transport-rest/src/main/java/ptc/springframework/publictransportrest/service/TicketService.java
package ptc.springframework.publictransportrest.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import ptc.springframework.publictransportrest.model.TicketType;
import ptc.springframework.publictransportrest.repository.TicketRepository;

import java.util.List;

@Service
public class TicketService {

    @Autowired
    TicketRepository ticketRepository;

    public List<TicketType> getTicketTypes() {
        return ticketRepository.findAll();
    }
}
